19: Server & Map Editor updated, server now replicates player positions and sends world data

+++ code/urban_tile.h
@@ -18,7 +18,7 @@ struct TileChunkPosition
 struct TileChunk
 {
     // TODO(bSalmon): Tile structure
-    u32 *tiles;
+    u8 *tiles;
 };
 
 struct TileMap
